Move function 'showDataNotLoadedError' from 'util.js' to 'page.js'
