+++ js/main.js
@@ -1,6 +1,5 @@
 import { GET_ANNOUNCEMENTS_URL } from './config.js';
-import { showDataNotLoadedError } from './util.js';
-import { activatePage, deactivatePage } from './page.js';
+import { activatePage, deactivatePage, showDataNotLoadedError } from './page.js';
 import './map.js';
 import {interactiveMap, createMarkers} from './map.js';
 import './form.js';
