+++ js/page.js
@@ -8,4 +8,12 @@ const deactivatePage = () => {
   formClassNames.forEach((className) => deactivateForm(className));
 };
 
-export {activatePage, deactivatePage};
+const showDataNotLoadedError = (errorMessage) => {
+  const mapElement = document.querySelector('.map');
+  const errorElement = document.createElement('div');
+  errorElement.classList.add('map__error');
+  errorElement.textContent = errorMessage;
+  mapElement.appendChild(errorElement);
+};
+
+export {activatePage, deactivatePage, showDataNotLoadedError};
